Share single Spring context across the transactional integration tests
