+++ sbs/axelix-spring-boot-2-starter/src/test/java/com/axelixlabs/axelix/sbs/spring/core/transactions/ProxyingDataSourceBeanPostProcessorTest.java
@@ -17,15 +17,10 @@
  */
 package com.axelixlabs.axelix.sbs.spring.core.transactions;
 
-import javax.sql.DataSource;
-
 import org.junit.jupiter.api.Test;
-
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Import;
+
+import javax.sql.DataSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -34,9 +29,7 @@ import static org.assertj.core.api.Assertions.assertThat;
  *
  * @author Sergey Cherkasov
  */
-@SpringBootTest
-@Import(ProxyingDataSourceBeanPostProcessorTest.ProxyingDataSourceBeanPostProcessorTestConfiguration.class)
-class ProxyingDataSourceBeanPostProcessorTest {
+class ProxyingDataSourceBeanPostProcessorTest extends AbstractTransactionalIntegrationTest {
 
     @Autowired
     private DataSource dataSource;
@@ -64,20 +57,5 @@ class ProxyingDataSourceBeanPostProcessorTest {
         Object result = subject.postProcessAfterInitialization(alreadyProxied, "dataSource");
 
         assertThat(result).isSameAs(alreadyProxied);
-    }
-
-    @TestConfiguration
-    static class ProxyingDataSourceBeanPostProcessorTestConfiguration {
-
-        @Bean
-        public QueriesRecorder queriesRecorder() {
-            return new DefaultQueriesRecorder();
-        }
-
-        @Bean
-        public ProxyingDataSourceBeanPostProcessor proxyingDataSourceBeanPostProcessor(
-                QueriesRecorder queriesRecorder) {
-            return new ProxyingDataSourceBeanPostProcessor(queriesRecorder);
-        }
     }
 }
