+++ sbs/axelix-spring-boot-2-starter/src/test/java/com/axelixlabs/axelix/sbs/spring/core/transactions/TransactionMonitoringEndpointTest.java
@@ -224,11 +224,11 @@ class TransactionMonitoringEndpointTest extends AbstractEndpointIntegrationTest 
                                 + "      \"startTimestampMs\" : \"#{json-unit.ignore}\",\n"
                                 + "      \"endTimestampMs\" : \"#{json-unit.ignore}\",\n"
                                 + "      \"queries\" : [ {\n"
-                                + "        \"sql\" : \"select transactio0_.id as id1_0_0_, transactio0_.last_name as last_nam2_0_0_ from owner transactio0_ where transactio0_.id=?\",\n"
+                                + "        \"sql\" : \"select transactio0_.id as id1_1_0_, transactio0_.last_name as last_nam2_1_0_ from owner transactio0_ where transactio0_.id=?\",\n"
                                 + "        \"startTimestampMs\" : \"#{json-unit.ignore}\",\n"
                                 + "        \"endTimestampMs\" : \"#{json-unit.ignore}\"\n"
                                 + "      }, {\n"
-                                + "        \"sql\" : \"select pets0_.owner_id as owner_id3_1_0_, pets0_.id as id1_1_0_, pets0_.id as id1_1_1_, pets0_.name as name2_1_1_, pets0_.owner_id as owner_id3_1_1_ from pet pets0_ where pets0_.owner_id=?\",\n"
+                                + "        \"sql\" : \"select pets0_.owner_id as owner_id3_2_0_, pets0_.id as id1_2_0_, pets0_.id as id1_2_1_, pets0_.name as name2_2_1_, pets0_.owner_id as owner_id3_2_1_ from pet pets0_ where pets0_.owner_id=?\",\n"
                                 + "        \"startTimestampMs\" : \"#{json-unit.ignore}\",\n"
                                 + "        \"endTimestampMs\" : \"#{json-unit.ignore}\"\n"
                                 + "      } ]\n"
@@ -282,7 +282,7 @@ class TransactionMonitoringEndpointTest extends AbstractEndpointIntegrationTest 
                                 + "      \"startTimestampMs\" : \"#{json-unit.ignore}\",\n"
                                 + "      \"endTimestampMs\" : \"#{json-unit.ignore}\",\n"
                                 + "      \"queries\" : [ {\n"
-                                + "        \"sql\" : \"select transactio0_.id as id1_0_0_, transactio0_.last_name as last_nam2_0_0_ from owner transactio0_ where transactio0_.id=?\",\n"
+                                + "        \"sql\" : \"select transactio0_.id as id1_1_0_, transactio0_.last_name as last_nam2_1_0_ from owner transactio0_ where transactio0_.id=?\",\n"
                                 + "        \"startTimestampMs\" : \"#{json-unit.ignore}\",\n"
                                 + "        \"endTimestampMs\" : \"#{json-unit.ignore}\"\n"
                                 + "      }, {\n"
@@ -338,7 +338,7 @@ class TransactionMonitoringEndpointTest extends AbstractEndpointIntegrationTest 
                                 + "      \"startTimestampMs\" : \"#{json-unit.ignore}\",\n"
                                 + "      \"endTimestampMs\" : \"#{json-unit.ignore}\",\n"
                                 + "      \"queries\" : [ {\n"
-                                + "        \"sql\" : \"select transactio0_.id as id1_0_, transactio0_.last_name as last_nam2_0_ from owner transactio0_ where transactio0_.last_name=?\",\n"
+                                + "        \"sql\" : \"select transactio0_.id as id1_1_, transactio0_.last_name as last_nam2_1_ from owner transactio0_ where transactio0_.last_name=?\",\n"
                                 + "        \"startTimestampMs\" : \"#{json-unit.ignore}\",\n"
                                 + "        \"endTimestampMs\" : \"#{json-unit.ignore}\"\n"
                                 + "      } ]\n"
